scripts/whatsapp-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymysql==1.1.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.12
